[TASK] return (C)TpsaOr(Double|Complex) as python base type
C++ wrapper side should use gtpsa::python::to_pyobject
Visitors to gtpsa_or_double vist const object
